Added kill process tree extension
