+++ UI/ExistingProcessesWindow.xaml.cs
@@ -11,6 +11,7 @@ using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 using System.Diagnostics;
 using System.Threading;
+using GolemUI.Extensions;
 
 namespace GolemUI
 {
@@ -85,15 +86,15 @@ namespace GolemUI
 
             foreach (var yagnaProcess in yagnaProcesses)
             {
-                yagnaProcess.Kill(/*entireProcessTree: true*/);
+                yagnaProcess.Kill(entireProcessTree: true);
             }
             foreach (var providerProcess in providerProcesses)
             {
-                providerProcess.Kill(/*entireProcessTree: true*/);
+                providerProcess.Kill(entireProcessTree: true);
             }
             foreach (var claymoreProcess in claymoreProcesses)
             {
-                claymoreProcess.Kill(/*entireProcessTree: true*/);
+                claymoreProcess.Kill(entireProcessTree: true);
             }
         }
